test: split test stack in fixtures/integration/unit
Context: I wanted to add unit tests around flatten_browser.js and
thought "Well, we have tests, but it's a mess".

I do think that now it's easier to add tests, let me know.

I also fixed race conditions in capabilities test, we were asking for
impossible platform + browser. Now stable tests.

cc @rase @defunctzombie
