+++ package.json
@@ -42,10 +42,11 @@
   },
   "devDependencies": {
     "after": "~0.8.1",
+    "bulk-require": "0.2.1",
     "mocha": "~1.16.2",
     "phantomjs": "1.9.12",
-    "through2": "0.6.3",
-    "tape": "3.5.0"
+    "tape": "3.5.0",
+    "through2": "0.6.3"
   },
   "repository": {
     "type": "git",
