+++ test/fixtures/builder-browserify/bar-to-baz-plugin.js
