+++ test/fixtures/builder-browserify/baz-to-qux-transform.js
