+++ test/fixtures/builder-browserify/entry.js
