+++ test/fixtures/builder-browserify/foo-to-bar-transform.js
